+++ docs/changelog/87978.yaml
@@ -1,0 +1,5 @@
+pr: 87978
+summary: Improve trained model stats API performance
+area: Machine Learning
+type: bug
+issues: []
